revert legacy API Key deprecation
This is to keep the legacy api endpoint alive for a little longer so
we can do a proper migration from Legacy API Keys to service accounts

This is a squashed revert of 2 commits. Making a single commit for
easier patching until we migrate the keys.

commit 6a19278f2337c6668b6edd9f970bcc24a55f53aa. (return 302 redirect
informing this endpoint's deprecation)
commit ebd17adecf22fa5daa03f4d907e2b03f6627d54f. (return 410 Gone for
/auth/keys endpoint)